+++ src/main/java/com/example/demo/entity/AuditLog.java
@@ -15,7 +15,7 @@ public class AuditLog {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "log_id")
-    private Integer logId;
+    private Long logId;
 
     @ManyToOne
     @JoinColumn(name = "user_id")
@@ -25,7 +25,7 @@ public class AuditLog {
     private String entityType;
 
     @Column(name = "entity_id")
-    private Integer entityId;
+    private Long entityId;
 
     private String action;
